Added printf to Python/ceval.c
We're printing each time an opcode is set as a PyTryBlock b_type
attribute. The point is to see what opcodes are set when we run various operations in the CLI to the interpreter.

There is also a comment in Lib/opcode.py to illustrate the opcodes we've found to be valid PyTryBlock b_types.